TrackSpend Front-End
For Sprint 1 of the TrackSpend application development, our focus was on establishing the foundational backend infrastructure and developing the core informational and user authentication interfaces. This included the creation of the home page, contact page, about page, and privacy policy page to provide essential user information and context. Additionally, we implemented the fundamental user management functionalities through the development of both login and registration pages.